Module calypso_base::streams

Utilities for handling streams of data

Modules

generic
Generic streams

str
String streams

wrapper
Streams that wrap iterators

Structs

Streamed
A generic stream.

StringStream
A stream emitting tuples of byte locations and characters from a string slice.

Traits

Stream
A trait for streamed data. For a stream, EOF means either the end of file / input if this stream is backed by a file or string, or the end of the stream.