Trait calypso_base::streams::Stream

pub trait Stream: Iterator {

    fn is_at_end(&self) -> bool;

    fn peek(&self) -> Option<&<Self as Iterator>::Item>;

    fn peek2(&self) -> Option<&<Self as Iterator>::Item>;

    fn peek3(&self) -> Option<&<Self as Iterator>::Item>;

    fn prev(&self) -> Option<&<Self as Iterator>::Item>;

    fn next_if(
        &mut self, 
        func: impl FnOnce(&<Self as Iterator>::Item) -> bool
    ) -> Option<<Self as Iterator>::Item> { ... }

    fn next_if_eq<R>(
        &mut self, 
        expected: &R
    ) -> Option<<Self as Iterator>::Item>
    where
        R: ?Sized,
        <Self as Iterator>::Item: PartialEq<R>,
    { ... }

    fn gorge_while(
        &mut self, 
        func: impl FnMut(&<Self as Iterator>::Item, usize) -> bool
    ) -> usize { ... }

    fn gorge_while_eq<R>(&mut self, expected: &R) -> usize
    where
        R: ?Sized,
        <Self as Iterator>::Item: PartialEq<R>,
    { ... }

    fn peek_eq<R>(&self, expected: &R) -> Option<bool>
    where
        R: ?Sized,
        <Self as Iterator>::Item: PartialEq<R>,
    { ... }

    fn peek2_eq<R>(&self, expected: &R) -> Option<bool>
    where
        R: ?Sized,
        <Self as Iterator>::Item: PartialEq<R>,
    { ... }

    fn peek3_eq<R>(&self, expected: &R) -> Option<bool>
    where
        R: ?Sized,
        <Self as Iterator>::Item: PartialEq<R>,
    { ... }

    fn prev_eq<R>(&self, expected: &R) -> Option<bool>
    where
        R: ?Sized,
        <Self as Iterator>::Item: PartialEq<R>,
    { ... }

    fn peek_cond(
        &self, 
        func: impl FnOnce(&<Self as Iterator>::Item) -> bool
    ) -> Option<bool> { ... }

    fn peek2_cond(
        &self, 
        func: impl FnOnce(&<Self as Iterator>::Item) -> bool
    ) -> Option<bool> { ... }

    fn peek3_cond(
        &self, 
        func: impl FnOnce(&<Self as Iterator>::Item) -> bool
    ) -> Option<bool> { ... }

    fn prev_cond(
        &self, 
        func: impl FnOnce(&<Self as Iterator>::Item) -> bool
    ) -> Option<bool> { ... }

}

A trait for streamed data. For a stream, EOF means either the end of file / input if this stream is backed by a file or string, or the end of the stream.

Required methods

fn is_at_end(&self) -> bool

This stream can no longer produce elements if this returns true

fn peek(&self) -> Option<&<Self as Iterator>::Item>

Peek at the value one element ahead, without consuming it.

fn peek2(&self) -> Option<&<Self as Iterator>::Item>

Peek at the value two element ahead, without consuming anything.

fn peek3(&self) -> Option<&<Self as Iterator>::Item>

Peek at the value three element ahead, without consuming anything.

fn prev(&self) -> Option<&<Self as Iterator>::Item>

Peek at the value one element behind, without moving backwards.

Provided methods

fn next_if(
&mut self,
func: impl FnOnce(&<Self as Iterator>::Item) -> bool
) -> Option<<Self as Iterator>::Item>

Consume the next element if the condition is true.

If func returns true for the element, consume and return it. Otherwise, return None.

fn next_if_eq<R>(&mut self, expected: &R) -> Option<<Self as Iterator>::Item> where
R: ?Sized,
<Self as Iterator>::Item: PartialEq<R>,

Consume the next element if it is equal to expected.

fn gorge_while(
&mut self,
func: impl FnMut(&<Self as Iterator>::Item, usize) -> bool
) -> usize

Keep consuming elements while the condition is true and there are remaining elements. Returns the number of elements consumed.

If func returns true for the element, consume it. Repeat until func returns false for an element.

func takes in the element first, then the number of elements consumed so far.

fn gorge_while_eq<R>(&mut self, expected: &R) -> usize where
R: ?Sized,
<Self as Iterator>::Item: PartialEq<R>,

Keep consuming elements while the element is equal to expected and there are remaining elements. Returns the number of elements consumed.

fn peek_eq<R>(&self, expected: &R) -> Option<bool> where
R: ?Sized,
<Self as Iterator>::Item: PartialEq<R>,

Returns Some(true) if the value one element ahead is equal to expected. Otherwise, return Some(false). In neither case should any values be consumed. On EOF, return None.

fn peek2_eq<R>(&self, expected: &R) -> Option<bool> where
R: ?Sized,
<Self as Iterator>::Item: PartialEq<R>,

Returns Some(true) if the value two elements ahead is equal to expected. Otherwise, return Some(false). In neither case should any values be consumed. On EOF, return None.

fn peek3_eq<R>(&self, expected: &R) -> Option<bool> where
R: ?Sized,
<Self as Iterator>::Item: PartialEq<R>,

Returns Some(true) if the value three elements ahead is equal to expected. Otherwise, return Some(false). In neither case should any values be consumed. On EOF, return None.

fn prev_eq<R>(&self, expected: &R) -> Option<bool> where
R: ?Sized,
<Self as Iterator>::Item: PartialEq<R>,

Returns Some(true) if the value one element behind is equal to expected. Otherwise, return Some(false). In neither case should any values be consumed. On EOF, return None.

fn peek_cond(
&self,
func: impl FnOnce(&<Self as Iterator>::Item) -> bool
) -> Option<bool>

Returns the value of the condition, wrapped in an Option. In neither case should any values be consumed. On EOF, return None.

func takes in the value one element ahead.

fn peek2_cond(
&self,
func: impl FnOnce(&<Self as Iterator>::Item) -> bool
) -> Option<bool>

Returns the value of the condition, wrapped in an Option. In neither case should any values be consumed. On EOF, return None.

func takes in the value two elements ahead.

fn peek3_cond(
&self,
func: impl FnOnce(&<Self as Iterator>::Item) -> bool
) -> Option<bool>

Returns the value of the condition, wrapped in an Option. In neither case should any values be consumed. On EOF, return None.

func takes in the value three elements ahead.

fn prev_cond(
&self,
func: impl FnOnce(&<Self as Iterator>::Item) -> bool
) -> Option<bool>

Returns the value of the condition, wrapped in an Option. In neither case should any values be consumed. On EOF, return None.

func takes in the value one element behind.