Module calypso_base::symbol::kw

Keywords

Structs

DO

Do (do)

END

End (end)

FALSE

False (false)

IN

In (in)

LET

Let (let)

MUT

Mut (mut)

TRUE

True (true)

Enums

Keyword

An enum containing all of the statically interned values in this module.

Functions

init 🔒

Initialize all of the symbols in this module.

is

Check if the given symbol is one of the statically interned members in this module.